refactor into import class
